+++ tencentcloud/extension_ssl.go
@@ -12,6 +12,36 @@ const (
 	SSL_CERT_TYPE_CA     = "CA"
 )
 
+const (
+	CsrTypeOnline = "online"
+	CsrTypeParse  = "parse"
+)
+
+const (
+	DNSAuto = "DNS_AUTO"
+	DNS     = "DNS"
+	File    = "FILE"
+)
+
+const (
+	InvalidParam          = "FailedOperation.InvalidParam"
+	CertificateNotFound   = "FailedOperation.CertificateNotFound"
+	InvalidParameter      = "InvalidParameter"
+	InvalidParameterValue = "InvalidParameterValue"
+	CertificateInvalid    = "FailedOperation.CertificateInvalid"
+)
+
+var CsrTypeArr = []string{
+	CsrTypeOnline,
+	CsrTypeParse,
+}
+
+var VerifyType = []string{
+	DNSAuto,
+	DNS,
+	File,
+}
+
 var SSL_CERT_TYPE = []string{
 	SSL_CERT_TYPE_SERVER,
 	SSL_CERT_TYPE_CA,