+++ tencentcloud/resource_tc_ccn_bandwidth_limit.go
@@ -3,6 +3,8 @@ Provides a resource to limit CCN bandwidth.
 
 Example Usage
 
+Set the upper limit of regional outbound bandwidth
+
 ```hcl
 variable "other_region1" {
   default = "ap-shanghai"
@@ -20,6 +22,33 @@ resource "tencentcloud_ccn_bandwidth_limit" "limit1" {
   bandwidth_limit = 500
 }
 ```
+
+Set the upper limit between regions
+
+```hcl
+variable "other_region1" {
+  default = "ap-shanghai"
+}
+
+variable "other_region2" {
+  default = "ap-nanjing"
+}
+
+resource tencentcloud_ccn main {
+  name                 = "ci-temp-test-ccn"
+  description          = "ci-temp-test-ccn-des"
+  qos                  = "AG"
+  bandwidth_limit_type = "INTER_REGION_LIMIT"
+}
+
+resource tencentcloud_ccn_bandwidth_limit limit1 {
+  ccn_id          = tencentcloud_ccn.main.id
+  region          = var.other_region1
+  dst_region      = var.other_region2
+  bandwidth_limit = 100
+}
+```
+
 */
 package tencentcloud
 
@@ -57,6 +86,13 @@ func resourceTencentCloudCcnBandwidthLimit() *schema.Resource {
 				Computed:    true,
 				Description: "Limitation of bandwidth.",
 			},
+			"dst_region": {
+				Type:     schema.TypeString,
+				Optional: true,
+				ForceNew: true,
+				Description: "Destination area restriction. If the `CCN` rate limit type is `OUTER_REGION_LIMIT`, " +
+					"this value does not need to be set.",
+			},
 		},
 	}
 }
@@ -81,13 +117,19 @@ func resourceTencentCloudCcnBandwidthLimitCreate(d *schema.ResourceData, meta in
 	if has == 0 {
 		return fmt.Errorf("ccn[%s] doesn't exist", ccnId)
 	}
-
 	id := fmt.Sprintf("%s#%s", ccnId, region)
-
-	if limitTemp, ok := d.GetOk("bandwidth_limit"); ok {
-		if err := service.SetCcnRegionBandwidthLimits(ctx, ccnId, region, int64(limitTemp.(int))); err != nil {
-			return err
-		}
+	var (
+		dstRegion string
+		limit     int64
+	)
+	if v, ok := d.GetOkExists("dst_region"); ok {
+		dstRegion = v.(string)
+	}
+	if v, ok := d.GetOk("bandwidth_limit"); ok {
+		limit = int64(v.(int))
+	}
+	if err := service.SetCcnRegionBandwidthLimits(ctx, ccnId, region, dstRegion, limit); err != nil {
+		return err
 	}
 	d.SetId(id)
 
@@ -113,13 +155,16 @@ func resourceTencentCloudCcnBandwidthLimitUpdate(d *schema.ResourceData, meta in
 	if has == 0 {
 		return fmt.Errorf("ccn[%s] doesn't exist", ccnId)
 	}
-
-	if limitTemp, ok := d.GetOk("bandwidth_limit"); ok {
-		if err := service.SetCcnRegionBandwidthLimits(ctx, ccnId, region, int64(limitTemp.(int))); err != nil {
+	if d.HasChange("bandwidth_limit") {
+		var limitTemp int64
+		if v, ok := d.GetOk("bandwidth_limit"); ok {
+			limitTemp = int64(v.(int))
+		}
+		_, dstRegion := d.GetChange("dst_region")
+		if err := service.SetCcnRegionBandwidthLimits(ctx, ccnId, region, dstRegion.(string), limitTemp); err != nil {
 			return err
 		}
 	}
-
 	return resourceTencentCloudCcnBandwidthLimitRead(d, meta)
 }
 
@@ -135,10 +180,12 @@ func resourceTencentCloudCcnBandwidthLimitRead(d *schema.ResourceData, meta inte
 	var (
 		ccnId     = d.Get("ccn_id").(string)
 		region    = d.Get("region").(string)
+		dstRegion = d.Get("dst_region").(string)
 		onlineHas = true
+		info      CcnBasicInfo
 	)
 	err := resource.Retry(readRetryTimeout, func() *resource.RetryError {
-		_, has, e := service.DescribeCcn(ctx, ccnId)
+		infoTmp, has, e := service.DescribeCcn(ctx, ccnId)
 		if e != nil {
 			return retryError(e)
 		}
@@ -148,6 +195,7 @@ func resourceTencentCloudCcnBandwidthLimitRead(d *schema.ResourceData, meta inte
 			onlineHas = false
 			return nil
 		}
+		info = infoTmp
 		return nil
 	})
 	if err != nil {
@@ -157,11 +205,12 @@ func resourceTencentCloudCcnBandwidthLimitRead(d *schema.ResourceData, meta inte
 		return nil
 	}
 	err = resource.Retry(readRetryTimeout, func() *resource.RetryError {
-		bandwidth, e := service.DescribeCcnRegionBandwidthLimit(ctx, ccnId, region)
+		bandwidth, e := service.GetCcnRegionBandwidthLimit(ctx, ccnId, region, dstRegion, info.bandWithLimitType)
 		if e != nil {
 			return retryError(e)
 		}
 		_ = d.Set("bandwidth_limit", bandwidth)
+		_ = d.Set("dst_region", dstRegion)
 		return nil
 	})
 	if err != nil {